feat: create special dtos for boarding-pass controller
