+++ src/main/java/org/vr61v/mappers/BoardingPassMapper.java
@@ -2,14 +2,17 @@ package org.vr61v.mappers;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import org.vr61v.dtos.BoardingPassDto;
+import org.vr61v.dtos.boardingPass.BoardingPassRequestDto;
+import org.vr61v.dtos.boardingPass.BoardingPassResponseDto;
 import org.vr61v.entities.BoardingPass;
 
 @Mapper(componentModel = "spring")
-public interface BoardingPassMapper extends BaseMapper<BoardingPass, BoardingPassDto> {
+public interface BoardingPassMapper {
 
     @Mapping(target = "ticketNo", source = "entity.id.ticket.ticketNo")
     @Mapping(target = "flightId", source = "entity.id.flight.flightId")
-    BoardingPassDto toDto(BoardingPass entity);
+    BoardingPassResponseDto toDto(BoardingPass entity);
+
+    BoardingPass toEntity(BoardingPassRequestDto dto);
 
 }
